requirements.txt
certifi==2018.10.15
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.5.0
Django==2.1.3
django-oauth-toolkit==1.2.0
django-registration==3.0
idna==2.7
image==1.5.27
oauthlib==2.1.0
Pillow==5.3.0
PyJWT==1.6.4
python3-openid==3.1.0
pytube==9.3.0
pytz==2018.5
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-core==2.0.0
urllib3==1.24.1
